requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
emoji==1.4.1
flake8==3.9.2
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
PySocks==1.7.1
python-telegram-bot[socks]==12.6.1
pytz==2021.1
six==1.16.0
tornado==6.1
tzlocal==2.1
